fix: get handler is avaliable and patient's data back in create handler

+++ server/internal/entities/patient.go
@@ -5,6 +5,6 @@ type Patient struct {
 	Name        string `json:"name"`
 	LastName    string `json:"last_name"`
 	DateOfBirth string `json:"date_of_birth"`
-	BloodType   uint8  `json:"blood_type"`
+	BloodType   uint   `json:"blood_type"`
 	RhFactor    string `json:"rh_factor"`
 }
